refactor components out into seperate files
