--- app/javascript/pages/MyEvents/main.css
@@ -1,166 +1,0 @@
-.muiTextField hr {
-  display: none;
-}
-
-.actionBar {
-  display: flex;
-  justify-content: flex-end;
-  position: absolute;
-  right: 0;
-  bottom: 0;
-}
-
-.createAction {
-  border: 1px solid #ddd;
-  border-radius: 4px;
-  background: none;
-  padding: 8px 20px;
-}
-
-.createAction:active {
-  border-color: #ddd;
-  background: #f5f5f5;
-}
-
-.table {
-  width: 100%;
-  border-collapse: collapse;
-}
-
-.table th {
-  border-bottom: 2px solid #eee;
-  text-align: left;
-  padding: 15px 5px;
-}
-
-.table td {
-  padding: 5px 10px;
-  border-bottom: 1px solid #eee;
-}
-
-.table tr:last-of-type td {
-  border-bottom: none;
-}
-
-.table tr td {
-  max-width: 200px;
-  white-space: nowrap;
-  text-overflow: ellipsis;
-  overflow: hidden;
-}
-
-.actionColumn {
-  max-width: 200px;
-  white-space: nowrap;
-}
-
-.btn {
-  border: none;
-  background: none;
-  padding: 5px 15px;
-}
-
-.cancelBtn {
-  color: #ff542e;
-}
-
-.confirmBtn {
-  color: #08c;
-}
-
-.deleteBtn {
-  color: #a93a38;
-}
-
-.input {
-  font-size: 16px;
-  border: 1px solid #eee;
-  border-radius: 4px;
-  padding: 5px 10px;
-  min-width: 250px;
-  background: none;
-  min-height: 20px;
-  color: #555;
-}
-
-select.input {
-  height: 30px; /* selects don't honor padding, account for it in the height */
-}
-
-.errorMsg {
-  color: #ff542e;
-  font-size: smaller;
-  font-weight: normal;
-}
-
-.form {
-  display: flex;
-  flex-flow: column nowrap;
-}
-
-.form label {
-  color: #333;
-  font-weight: bold;
-  display: block;
-  padding-bottom: 5px;
-}
-
-.form input,
-textarea {
-  width: calc(100% - 20px);
-}
-
-.form textarea {
-  height: auto;
-}
-
-.form select {
-  width: 100%;
-}
-
-.form .input {
-  border-color: #ddd;
-}
-
-.formSpace {
-  height: 20px;
-}
-
-.personalHeader {
-  position: relative;
-}
-
-.row {
-  display: flex;
-  flex-flow: row nowrap;
-  justify-content: space-between;
-}
-
-.column {
-  width: 47%;
-}
-
-.dialogBottomBar {
-  padding-top: 20px;
-  align-items: center;
-}
-
-.dialogActions {
-  justify-content: flex-end;
-}
-
-.eventsTable > div {
-  border: none;
-}
-
-.eventsTable > div > div:nth-child(1) > div:nth-child(1) {
-  box-shadow: inset 0 -2px #eee;
-}
-
-.eventsTable {
-  margin-bottom: 20px;
-}
-
-.noEventsMessage {
-  font-style: italic;
-}
